fix: 日記編集時、新たな画像を選択しなかった際に既存の画像が削除されないように修正

+++ app/Services/ImageUploadService.php
@@ -9,7 +9,10 @@ class ImageUploadService
 {
     public function uploadImage($request, $post)
     {
-        $this->deleteImage($post->image);
+        if ($request->delete_image == '1' || $request->image) {
+            $this->deleteImage($post->image);
+            $post->image = null;
+        }
 
         if ($request->hasFile('image')) {
 
@@ -35,8 +38,6 @@ class ImageUploadService
             }
 
             $resizedImage->destroy();
-        } else {
-            $post->image = null;
         }
     }
